requirements.txt
fastapi==0.116.1
pydantic==2.11.7
starlette==0.47.2

uvicorn~=0.35.0
python-dotenv~=1.1.0
pydantic-settings~=2.10.1
slowapi~=0.1.9
asyncpg~=0.30.0
pytest~=8.3.5
SQLAlchemy~=2.0.43

passlib~=1.7.4
uliweb-alembic~=0.6.9
asyncio~=4.0.0
argparse~=1.4.0
bcrypt~=4.3.0
greenlet~=3.2.3